Implemented clickLinkWithTextAfterText.
The use case is a test that wanted to click a link with text "...", when
several identical links appeared on the page. At first we tried to use the
indexed form of clickLinkWithText, but that turned out to be fragile. Instead
we realized we could distinguish between them by a label that preceded each
link on the page.

(Didn't write assert/assertNot yet.)


git-svn-id: https://jwebunit.svn.sourceforge.net/svnroot/jwebunit/trunk@182 aee3fcf7-a411-0410-a948-fdebeb6d3bc7